# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Microfluidic API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Result",
        "operationId": "get_job_result_api_v1_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/artifacts": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Artifacts",
        "operationId": "list_artifacts_api_v1_jobs__job_id__artifacts_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/download/{name}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Download Artifact",
        "operationId": "download_artifact_api_v1_jobs__job_id__download__name__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoundarySpec": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "BoundarySpec"
      },
...
      "GeometrySpec": {
        "properties": {
          "width": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "width",
          "height"
        ],
        "title": "GeometrySpec"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobSpec": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeometrySpec"
          },
          "material": {
            "$ref": "#/components/schemas/MaterialSpec"
          },
          "boundaries": {
            "items": {
              "$ref": "#/components/schemas/BoundarySpec"
            },
            "type": "array",
            "title": "Boundaries"
          },
          "solve_transport": {
            "type": "boolean",
            "title": "Solve Transport",
            "default": true
          },
          "geometry_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry Json"
          }
        },
        "type": "object",
        "required": [
          "name",
          "geometry",
          "material",
          "boundaries"
        ],
        "title": "JobSpec"
      },
      "JobStatus": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "JobStatus"
      },
      "MaterialSpec": {
        "properties": {
          "density": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Density"
          },
          "viscosity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Viscosity"
          },
          "diffusivity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Diffusivity"
          }
        },
        "type": "object",
        "required": [
          "density",
          "viscosity",
          "diffusivity"
        ],
        "title": "MaterialSpec"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}